# backend/api/openapi.yaml
openapi: "3.0.3"
info:
  title: SummitMate API
  description: 嘉明湖登山行程助手 - 後端 API
  version: 0.1.0
  contact:
    name: SummitMate
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /gear-library:
    get:
      summary: List Gear Library Items
      description: 取得使用者的個人裝備庫列表 (已封存的可透過參數過濾)
      operationId: listGearLibrary
      tags:
        - gear-library
      security:
        - bearerAuth: []
      parameters:
        - name: include_archived
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - name: page
          in: query
          description: 頁碼 (1-indexed)
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          description: 每頁筆數
          required: false
          schema:
            type: integer
            default: 50
        - name: search
          in: query
          description: 搜尋關鍵字
          required: false
          schema:
            type: string
      responses:
        "200":
          description: 成功回傳列表與分頁資訊
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GearLibraryPaginationResponse"
        "401":
          description: 未登入
    post:
      summary: Create Gear Library Item
      description: 新增個人裝備至庫中
      operationId: createGearLibraryItem
      tags:
        - gear-library
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GearLibraryItemRequest"
      responses:
        "201":
          description: 成功建立裝備
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GearLibraryItem"
        "400":
          description: 參數錯誤
        "401":
          description: 未登入
    put:
      summary: Replace All Gear Library Items
      description: 批量替換使用者的全部裝備庫 (離線同步用)
      operationId: replaceAllGearLibraryItems
      tags:
        - gear-library
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/GearLibraryItem"
      responses:
        "200":
          description: 批量替換成功
        "401":
          description: 未登入
  /gear-library/{itemId}:
    get:
      summary: Get Gear Library Item
      description: 取得單一裝備詳情
      operationId: getGearLibraryItem
      tags:
        - gear-library
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 成功取得裝備
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GearLibraryItem"
        "401":
          description: 未登入
        "403":
          description: 無權限存取
        "404":
          description: 找不到該裝備
    put:
      summary: Update Gear Library Item
      description: 更新個人裝備資料
      operationId: updateGearLibraryItem
      tags:
        - gear-library
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GearLibraryItemRequest"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GearLibraryItem"
        "401":
          description: 未登入
        "403":
          description: 無權限編輯
        "404":
          description: 找不到該裝備
    delete:
      summary: Delete Gear Library Item
      description: 刪除個人裝備 (支援實體刪除)
      operationId: deleteGearLibraryItem
      tags:
        - gear-library
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: 成功刪除
        "401":
          description: 未登入
        "403":
          description: 無權限編輯
        "404":
          description: 找不到該裝備
  /meal-library:
    get:
      summary: List Meal Library Items
      description: 取得使用者的個人食物庫列表
      operationId: listMealLibrary
      tags:
        - meal-library
      security:
        - bearerAuth: []
      parameters:
        - name: include_archived
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - name: page
          in: query
          description: 頁碼 (1-indexed)
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          description: 每頁筆數
          required: false
          schema:
            type: integer
            default: 50
        - name: search
          in: query
          description: 搜尋關鍵字
          required: false
          schema:
            type: string
      responses:
        "200":
          description: 成功回傳食物列表與分頁資訊
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealLibraryPaginationResponse"
        "401":
          description: 未登入
    post:
      summary: Create Meal Library Item
      description: 新增個人食物至庫中
      operationId: createMealLibraryItem
      tags:
        - meal-library
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MealLibraryItemRequest"
      responses:
        "201":
          description: 成功建立食物
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealLibraryItem"
        "400":
          description: 參數錯誤
        "401":
          description: 未登入
    put:
      summary: Replace All Meal Library Items
      description: 批量替換使用者的全部食物庫 (離線同步用)
      operationId: replaceAllMealLibraryItems
      tags:
        - meal-library
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/MealLibraryItem"
      responses:
        "200":
          description: 批量替換成功
        "401":
          description: 未登入
  /meal-library/{itemId}:
    get:
      summary: Get Meal Library Item
      description: 取得單一食物詳情
      operationId: getMealLibraryItem
      tags:
        - meal-library
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 成功取得食物
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealLibraryItem"
        "401":
          description: 未登入
        "403":
          description: 無權限存取
        "404":
          description: 找不到該食物
    put:
      summary: Update Meal Library Item
      description: 更新個人食物資料
      operationId: updateMealLibraryItem
      tags:
        - meal-library
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MealLibraryItemRequest"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealLibraryItem"
        "401":
          description: 未登入
        "403":
          description: 無權限編輯
        "404":
          description: 找不到該食物
    delete:
      summary: Delete Meal Library Item
      description: 刪除個人食物
      operationId: deleteMealLibraryItem
      tags:
        - meal-library
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: 成功刪除
        "401":
          description: 未登入
        "403":
          description: 無權限編輯
        "404":
          description: 找不到該食物
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    GearLibraryItem:
      type: object
      required:
        - id
        - user_id
        - name
        - weight
        - category
        - is_archived
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
        weight:
          type: number
          format: double
        category:
          type: string
        notes:
          type: string
          nullable: true
        is_archived:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    GearLibraryItemRequest:
      type: object
      required:
        - name
        - weight
        - category
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        weight:
          type: number
          format: double
          minimum: 0
        category:
          type: string
          maxLength: 50
        notes:
          type: string
          nullable: true
        is_archived:
          type: boolean
          default: false
    MealLibraryItem:
      type: object
      required:
        - id
        - user_id
        - name
        - weight
        - calories
        - is_archived
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
        weight:
          type: number
          format: double
        calories:
          type: number
          format: double
        notes:
          type: string
          nullable: true
        is_archived:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    MealLibraryItemRequest:
      type: object
      required:
        - name
        - weight
        - calories
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        weight:
          type: number
          format: double
          minimum: 0
        calories:
          type: number
          format: double
          minimum: 0
        notes:
          type: string
          nullable: true
        is_archived:
          type: boolean
          default: false
...
    PaginationMetadata:
      type: object
      required: [has_more, page, limit, total]
      properties:
        has_more:
          type: boolean
          description: 是否還有下一頁
        page:
          type: integer
          description: 目前頁碼 (1-indexed)
        limit:
          type: integer
          description: 每頁筆數
        total:
          type: integer
          description: 符合條件的總筆數
        next_cursor:
          type: string
          description: 下一頁的指標（cursor-based 相容用，可為 null）
          nullable: true
...
    GearLibraryPaginationResponse:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/GearLibraryItem"
        pagination:
          $ref: "#/components/schemas/PaginationMetadata"
    MealLibraryPaginationResponse:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MealLibraryItem"
        pagination:
          $ref: "#/components/schemas/PaginationMetadata"
...